Properly organize src/ include/
